refactor(loader): reorganize loader module into separate files
- Split monolithic loader.rs into modular structure
- Create loader/mod.rs as main module entry point
- Extract GitLoader into loader/git.rs with URL detection logic
- Extract LocalLoader into loader/local.rs for filesystem templates
- Create loader/interface.rs with TemplateLoader trait
- Update TemplateSource::from_string to load_from_string method
- Update cli.rs and tests to use new module structure
- Maintain backward compatibility for all loader functionality
