+++ src/loader/interface.rs
@@ -1,0 +1,11 @@
+use crate::error::Result;
+use std::path::PathBuf;
+
+/// Trait for loading templates from different sources.
+pub trait TemplateLoader {
+    /// Loads a template from the given source.
+    ///
+    /// # Returns
+    /// * `Result<PathBuf>` - Path to the loaded template
+    fn load(&self) -> Result<PathBuf>;
+}
